move useless_asref to its own module
